+++ src/main/java/net/cravencraft/epicparagliders/mixins/epicfight/skills/weaponinnate/SimpleSpecialAttackSkillMixin.java
@@ -1,4 +1,4 @@
-package net.cravencraft.epicparagliders.mixins.skills;
+package net.cravencraft.epicparagliders.mixins.epicfight.skills.weaponinnate;
 
 import net.cravencraft.epicparagliders.capabilities.PlayerMovementInterface;
 import net.cravencraft.epicparagliders.utils.MathUtils;
@@ -8,24 +8,22 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import tictim.paraglider.capabilities.PlayerMovement;
-import yesman.epicfight.skill.AirAttack;
 import yesman.epicfight.skill.Skill;
+import yesman.epicfight.skill.weaponinnate.SimpleWeaponInnateSkill;
 import yesman.epicfight.world.capabilities.entitypatch.player.ServerPlayerPatch;
 
-@Mixin(AirAttack.class)
-public abstract class AirAttackMixin extends Skill {
+@Mixin(SimpleWeaponInnateSkill.class)
+public abstract class SimpleSpecialAttackSkillMixin extends Skill {
 
-    public AirAttackMixin(Builder<? extends Skill> builder) {
+    public SimpleSpecialAttackSkillMixin(Builder<? extends Skill> builder) {
         super(builder);
     }
 
-    @Inject(method = "executeOnServer", at = @At("TAIL"), remap = false)
+    @Inject(method = "executeOnServer", at = @At("HEAD"), remap = false)
     private void getPlayerPatch(ServerPlayerPatch executer, FriendlyByteBuf args, CallbackInfo ci) {
         PlayerMovement playerMovement = PlayerMovement.of(executer.getOriginal());
-
         int specialAttackStaminaConsumption = MathUtils.getAttackStaminaCost(executer.getOriginal());
-
-        ((PlayerMovementInterface) playerMovement).setActionStaminaCostServerSide(specialAttackStaminaConsumption + 3);
+        ((PlayerMovementInterface) playerMovement).setActionStaminaCostServerSide(specialAttackStaminaConsumption);
         ((PlayerMovementInterface) playerMovement).isAttackingServerSide(true);
     }
 }
