+++ src/main/java/net/cravencraft/epicparagliders/mixins/paragliders/capabilities/PlayerMovementMixin.java
@@ -1,4 +1,4 @@
-package net.cravencraft.epicparagliders.mixins.capabilities;
+package net.cravencraft.epicparagliders.mixins.paragliders.capabilities;
 
 import net.cravencraft.epicparagliders.EpicParaglidersMod;
 import net.cravencraft.epicparagliders.capabilities.PlayerMovementInterface;
