+++ src/main/resources/epicparagliders.mixins.json
@@ -4,23 +4,23 @@
   "package": "net.cravencraft.epicparagliders.mixins",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
-    "capabilities.ItemCapabilityReloadListenerMixin",
-    "capabilities.ServerPlayerMovementMixin",
-    "capabilities.PlayerMovementMixin",
-    "skills.BasicAttackMixin",
-    "skills.AirAttackMixin",
-    "skills.weaponinnate.SimpleSpecialAttackSkillMixin",
-    "skills.guard.GuardSkillMixin",
-    "skills.dodge.DodgeSkillMixin",
-    "skills.guard.ParryingSkillMixin",
-    "skills.passive.StaminaPillagerSkillMixin",
-    "skills.passive.TechnicianSkillMixin",
-    "capabilities.PlayerPatchMixin"
+    "epicfight.capabilities.PlayerPatchMixin",
+    "epicfight.capabilities.ItemCapabilityReloadListenerMixin",
+    "epicfight.skills.BasicAttackMixin",
+    "epicfight.skills.AirAttackMixin",
+    "epicfight.skills.weaponinnate.SimpleSpecialAttackSkillMixin",
+    "epicfight.skills.guard.GuardSkillMixin",
+    "epicfight.skills.dodge.DodgeSkillMixin",
+    "epicfight.skills.guard.ParryingSkillMixin",
+    "epicfight.skills.passive.StaminaPillagerSkillMixin",
+    "epicfight.skills.passive.TechnicianSkillMixin",
+    "paragliders.capabilities.ServerPlayerMovementMixin",
+    "paragliders.capabilities.PlayerMovementMixin"
   ],
   "client": [
-    "client.ClientAnimatorMixin",
-    "client.InGameStaminaWheelRendererMixin",
-    "capabilities.ClientPlayerMovementMixin"
+    "epicfight.client.ClientAnimatorMixin",
+    "paragliders.client.InGameStaminaWheelRendererMixin",
+    "paragliders.capabilities.ClientPlayerMovementMixin"
   ],
   "injectors": {
     "defaultRequire": 1
